fixed the sizeof size_t, ptrdiff_t, intptr_t

+++ include/stdint.h
@@ -7,11 +7,11 @@
 
 // Types
 #if __x86_64__
-  typedef long int   int64_t;
+  typedef long long  intptr_t;
 #else
-  typedef long long  int64_t;
+  typedef long       intptr_t;
 #endif
-typedef long int     intptr_t;
+typedef long long    int64_t;
 typedef char         int8_t;
 typedef short        int16_t;
 typedef int          int32_t;
@@ -26,11 +26,11 @@ typedef int          int_least32_t;
 typedef int64_t      int_least64_t;
 
 #if __x86_64__
-  typedef unsigned long int   uint64_t;
+  typedef unsigned long long  uintptr_t;
 #else
-  typedef unsigned long long  uint64_t;
+  typedef unsigned long       uintptr_t;
 #endif
-typedef unsigned long int     uintptr_t;
+typedef unsigned long long    uint64_t;
 typedef unsigned char         uint8_t;
 typedef unsigned short        uint16_t;
 typedef unsigned int          uint32_t;
